// src/models/FamiliarData.ts
export interface FamiliarData {
  primerNombre: string;
  segundoNombre: string;
  primerApellido: string;
  segundoApellido: string;
  tipoDocumento: string;
  numeroDocumento: string;
  fechaNacimiento: string;
  genero: string;
  parentesco: string;
  eps: string;
  departamento: string;
  ciudad: string;
}